# docs/swagger/swagger.yaml
definitions:
...
  dto.CreatePaymentRequest:
    properties:
      amount:
        type: integer
      method:
        type: string
      purpose:
        type: string
      purpose_ref_id:
        type: string
    required:
    - amount
    - method
    - purpose
    type: object
...
info:
  contact: {}
  description: Payment service API (Gin)
  title: Modami Payment Service
  version: "1.0"
paths:
...
  /payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Create payment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Wrapped dto.PaymentResponse
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a payment transaction
      tags:
      - Payments
  /payments/{id}:
    get:
      parameters:
      - description: Payment ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.PaymentResponse
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get payment by ID
      tags:
      - Payments
  /payments/{id}/status:
    get:
      parameters:
      - description: Payment ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.PaymentStatusResponse
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get payment status by ID
      tags:
      - Payments
  /payments/history:
    get:
      parameters:
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped dto.PaymentHistoryResponse
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List payment history for current user
      tags:
      - Payments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"